Refactor yaml generation into the Metadata class itself
