+++ src/main/groovy/com/github/lookout/serviceartifact/Metadata.groovy
@@ -1,6 +1,10 @@
 package com.github.lookout.serviceartifact
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
+
+import com.github.lookout.serviceartifact.metadata.Data
 
 /**
  * Representation class for the metadata structure
@@ -23,15 +27,6 @@ class Metadata {
         String version
     }
 
-    class Data {
-        @JsonProperty
-        List<String> dependencies = []
-
-        @JsonProperty
-        List<String> migrations = []
-    }
-
-
     @JsonProperty
     Service service = new Service()
 
@@ -39,11 +34,19 @@ class Metadata {
     Component component = new Component()
 
     @JsonProperty
-    Data data = new Data()
+    Data data
 
     Metadata(String serviceName, String componentName, String componentVersion) {
         this.service.name = serviceName
         this.component.name = componentName
         this.component.version = componentVersion
+    }
+
+    /**
+     * @return This instance as a YAML String
+     */
+    String toYaml() {
+        ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+        return mapper.writeValueAsString(this)
     }
 }
